Add preact (with typescript)
Note that I have typescript preserving the jsx, leaving the
transpilation to babel, as suggested here:
https://preactjs.com/guide/v10/typescript/
But it turns out that typescript has a bug:
https://github.com/microsoft/TypeScript/issues/24731
that means it wasn't preserving the "h" import from preact, which is
needed for the jsx elements. But there seems to be a workaround of
setting "module": "ESNext" in tsconfig, and this seems to also require
setting "moduleResolution". Also, typescript requires that files with jsx
in be named .tsx.